backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pydantic[email]
python-multipart
langchain
langchain-community
langchain-openai
langchain-mongodb
pymongo[srv]
motor
pypdf
python-docx
tiktoken
python-dotenv
openai
sentence-transformers
python-jose
passlib
bcrypt
tenacity
httpx
pytest
pytest-asyncio
black
isort
flake8
mypy
gunicorn
loguru
motor